requirements.txt
click
Werkzeug==0.16.1
flask
# flask_cors
flask_bcrypt
flask-restplus
flask-injector
flask-jwt-extended
Flask-SQLAlchemy
mysql-connector-python
pycryptodome
pysaprk
